refactor(typescript): rewrite from esm to cjs
